added censor and repository implementation

+++ domain/src/main/java/com/harman/roomdbapp/domain/model/GravityValue.kt
@@ -3,13 +3,13 @@ package com.harman.roomdbapp.domain.model
 import kotlin.math.abs
 
 data class GravityValue(
-    val x: Int,
-    val y: Int,
-    val z: Int
+    val x: Float,
+    val y: Float,
+    val z: Float
 ) {
-    fun getFluctuation(previousValue: GravityValue): Int {
+    fun getFluctuation(previousValue: GravityValue): Float {
         return abs(previousValue.x - x) +
-                abs(previousValue.y - y) +
-                abs(previousValue.z - z)
+            abs(previousValue.y - y) +
+            abs(previousValue.z - z)
     }
 }
